refactor(kv-router): reorganize flat modules into scheduling/ and sequences/ subdirs
Move scheduling-related modules (config, types, selector, queue) into a
scheduling/ subdir and sequence-tracking modules (single, multi_worker) into
a sequences/ subdir. lib.rs re-exports preserve all old module paths so
external consumers (lib/llm, benches) need zero changes.
