+++ lib/kv-router/src/lib.rs
@@ -10,30 +10,34 @@ pub mod approx;
 #[cfg(feature = "bench")]
 pub mod bench_utils;
 pub mod concurrent_radix_tree;
-pub mod config;
 pub mod indexer;
-pub mod multi_worker_sequence;
 #[cfg(feature = "bench")]
 pub mod naive_indexers;
 pub mod nested_map;
 pub mod protocols;
-pub mod queue;
 pub mod radix_tree;
 pub mod scheduling;
-pub mod selector;
-pub mod sequence;
+pub mod sequences;
+
+// Backward-compat re-exports: preserve old module paths for external consumers
+pub use scheduling::config;
+pub use scheduling::queue;
+pub use scheduling::selector;
+pub use sequences::multi_worker as multi_worker_sequence;
+pub use sequences::single as sequence;
 
 #[cfg(test)]
 pub(crate) mod test_utils;
 
 // Re-export key types for convenience
-pub use concurrent_radix_tree::ConcurrentRadixTree;
-pub use config::{KvRouterConfig, RouterConfigOverride};
-pub use indexer::{MaybeError, SyncIndexer, ThreadPoolIndexer};
-pub use multi_worker_sequence::{
+pub use self::multi_worker_sequence::{
     ActiveSequencesMultiWorker, SequenceError, SequencePublisher, SequenceRequest,
     SequenceSubscriber,
 };
+pub use self::sequence::{ActiveSequences, RequestId};
+pub use concurrent_radix_tree::ConcurrentRadixTree;
+pub use config::{KvRouterConfig, RouterConfigOverride};
+pub use indexer::{MaybeError, SyncIndexer, ThreadPoolIndexer};
 #[cfg(feature = "bench")]
 pub use naive_indexers::{InvertedIndex, NaiveNestedMap};
 pub use nested_map::PositionalIndexer;
@@ -45,4 +49,3 @@ pub use queue::SchedulerQueue;
 pub use radix_tree::RadixTree;
 pub use scheduling::{KvSchedulerError, PotentialLoad, SchedulingRequest, SchedulingResponse};
 pub use selector::{DefaultWorkerSelector, WorkerSelector};
-pub use sequence::{ActiveSequences, RequestId};
