+++ lib/kv-router/src/scheduling/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod config;
+pub mod queue;
+pub mod selector;
+
+mod types;
+pub use types::*;
