+++ lib/kv-router/src/scheduling/types.rs
@@ -6,7 +6,7 @@ use std::collections::{HashMap, HashSet};
 use dynamo_tokens::SequenceHash;
 use serde::{Deserialize, Serialize};
 
-use crate::config::RouterConfigOverride;
+use super::config::RouterConfigOverride;
 use crate::protocols::{DpRank, OverlapScores, WorkerId, WorkerWithDpRank};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
